Order Management

+++ Restaurant.Services.OrderAPI/Models/OrderHeader.cs
@@ -10,7 +10,7 @@ namespace Restaurant.Services.OrderAPI.Models
         public string? UserId { get; set; }
         public string? CouponCode { get; set; }
         public double Discound { get; set; }
-        public double OrderTotal { get; set; }
+        public decimal OrderTotal { get; set; }
         public string? FirstName { get; set; }
         public string? LastName { get; set; }
         public string? Email { get; set; }
